+++ x/common/testutil/genesis/perp_genesis.go
@@ -37,7 +37,8 @@ var START_MARKETS = map[asset.Pair]perpammtypes.Market{
 		QuoteReserve:  sdk.NewDec(10 * common.TO_MICRO),
 		SqrtDepth:     common.MustSqrtDec(sdk.NewDec(10 * common.TO_MICRO * 10 * common.TO_MICRO)),
 		PegMultiplier: sdk.NewDec(6_000),
-		Bias:          sdk.ZeroDec(),
+		TotalLong:     sdk.ZeroDec(),
+		TotalShort:    sdk.ZeroDec(),
 		Config: perpammtypes.MarketConfig{
 			TradeLimitRatio:        sdk.MustNewDecFromStr("0.8"),
 			FluctuationLimitRatio:  sdk.MustNewDecFromStr("0.2"),
@@ -52,7 +53,8 @@ var START_MARKETS = map[asset.Pair]perpammtypes.Market{
 		QuoteReserve:  sdk.NewDec(10 * common.TO_MICRO),
 		SqrtDepth:     common.MustSqrtDec(sdk.NewDec(10 * common.TO_MICRO * 10 * common.TO_MICRO)),
 		PegMultiplier: sdk.NewDec(10),
-		Bias:          sdk.ZeroDec(),
+		TotalLong:     sdk.ZeroDec(),
+		TotalShort:    sdk.ZeroDec(),
 		Config: perpammtypes.MarketConfig{
 			TradeLimitRatio:        sdk.MustNewDecFromStr("0.8"),
 			FluctuationLimitRatio:  sdk.MustNewDecFromStr("0.2"),