+++ x/common/testutil/mock/perp_amm.go
@@ -16,7 +16,8 @@ func TestAMMDefault() *v2types.AMM {
 		QuoteReserve:    sdk.NewDec(1e12),
 		SqrtDepth:       sdk.NewDec(1e12),
 		PriceMultiplier: sdk.OneDec(),
-		Bias:            sdk.ZeroDec(),
+		TotalLong:       sdk.ZeroDec(),
+		TotalShort:      sdk.ZeroDec(),
 	}
 }
 
@@ -28,6 +29,7 @@ func TestAMM(sqrtK sdk.Dec, priceMultiplier sdk.Dec) *v2types.AMM {
 		QuoteReserve:    sqrtK,
 		SqrtDepth:       sqrtK,
 		PriceMultiplier: priceMultiplier,
-		Bias:            sdk.ZeroDec(),
+		TotalLong:       sdk.ZeroDec(),
+		TotalShort:      sdk.ZeroDec(),
 	}
 }