+++ x/perp/amm/abci.go
@@ -19,7 +19,6 @@ func EndBlocker(ctx sdk.Context, k keeper.Keeper) []abci.ValidatorUpdate {
 			pool.BaseReserve,
 			pool.QuoteReserve,
 			pool.PegMultiplier,
-			pool.Bias,
 			ctx.BlockTime(),
 		)
 		k.ReserveSnapshots.Insert(ctx, collections.Join(pool.Pair, ctx.BlockTime()), snapshot)