+++ x/wasm/binding/exec_perp_test.go
@@ -186,6 +186,17 @@ func (s *TestSuitePerpExecutor) DoPegShiftTest(pair asset.Pair) error {
 	return err
 }
 
+func (s *TestSuitePerpExecutor) DoDepthShiftTest(pair asset.Pair) error {
+	contractAddr := s.contractPerp
+	cwMsg := &cw_struct.DepthShift{
+		Pair:      pair.String(),
+		DepthMult: sdk.NewDec(420),
+	}
+
+	err := s.exec.DepthShift(cwMsg, contractAddr, s.ctx)
+	return err
+}
+
 func (s *TestSuitePerpExecutor) TestSadPaths_Nil() {
 	var err error
 
@@ -204,6 +215,10 @@ func (s *TestSuitePerpExecutor) TestSadPaths_Nil() {
 	err = s.exec.PegShift(
 		nil, sdk.AccAddress([]byte("contract")), s.ctx)
 	s.Error(err)
+
+	err = s.exec.DepthShift(
+		nil, sdk.AccAddress([]byte("contract")), s.ctx)
+	s.Error(err)
 }
 
 func (s *TestSuitePerpExecutor) TestSadPaths_InvalidPair() {
@@ -217,6 +232,7 @@ func (s *TestSuitePerpExecutor) TestSadPaths_InvalidPair() {
 		s.DoRemoveMarginTest(pair, margin),
 		s.DoClosePositionTest(pair),
 		s.DoPegShiftTest(pair),
+		s.DoDepthShiftTest(pair),
 	} {
 		s.Error(err)
 	}